+++ solr/__init__.py
@@ -10,3 +10,4 @@ from .mlt import MoreLikeThis
 from .suggest import Suggest
 from .extract import Extract
 from .response import SpellcheckResult
+from .response import GroupedResult, GroupField, Group